+++ tests/v2/test_02_models/test_taskgroup_constraints.py
@@ -298,3 +298,65 @@ async def test_taskgroup_activity_contraints(
                     )
                 )
         await db.commit()
+
+
+async def test_adding_task_group_activity(
+    MockCurrentUser, db, local_resource_profile_db
+):
+    resource, profile = local_resource_profile_db
+
+    async with MockCurrentUser(is_verified=True, profile_id=profile.id) as user:
+        # Create a new TaskGroupV2 associated to user.
+        task_group = TaskGroupV2(
+            user_id=user.id,
+            pkg_name="testing-tasks-mock",
+            version="0.1.4",
+            origin="pypi",
+            resource_id=resource.id,
+        )
+        db.add(task_group)
+        await db.commit()
+        await db.refresh(task_group)
+        task_group_id = task_group.id
+        # Create a TaskGroupActivityStatusV2 associated to the new TaskGroup.
+        task_group_activity_1 = TaskGroupActivityV2(
+            user_id=user.id,
+            taskgroupv2_id=task_group_id,
+            action=TaskGroupActivityAction.COLLECT,
+            status=TaskGroupActivityStatus.PENDING,
+            pkg_name="testing-tasks-mock",
+            version="0.1.4",
+        )
+        db.add(task_group_activity_1)
+        await db.commit()
+
+        # Create a new CollectionState associated to the same TaskGroup
+        # (this is NOT ALLOWED using the API).
+        db.add(
+            TaskGroupActivityV2(
+                user_id=user.id,
+                taskgroupv2_id=task_group_id,
+                action=TaskGroupActivityAction.COLLECT,
+                status=TaskGroupActivityStatus.PENDING,
+                pkg_name="testing-tasks-mock",
+                version="0.1.4",
+            )
+        )
+        with pytest.raises(
+            IntegrityError,
+            match="ix_taskgroupactivityv2_collect_unique_constraint",
+        ):
+            await db.commit()
+        await db.rollback()
+        db.add(
+            TaskGroupActivityV2(
+                user_id=user.id,
+                taskgroupv2_id=task_group_id,
+                # only action=COLLECT is constrained
+                action=TaskGroupActivityAction.DELETE,
+                status=TaskGroupActivityStatus.PENDING,
+                pkg_name="testing-tasks-mock",
+                version="0.1.4",
+            )
+        )
+        await db.commit()
